+++ R/data_tables.R
@@ -7,6 +7,7 @@
 #' - `time_at_sleep_onset`
 #' - `time_at_wakeup`
 #' - `time_in_bed`
+#' - `sleep_period`
 #' @returns A single-row dataframe summarizing session information.
 #' @importFrom rlang .data
 #' @export
@@ -20,7 +21,8 @@ get_sessions_summary <- function(sessions, col_names = NULL) {
       total_sessions = 0,
       mean_sleep_onset = NA,
       mean_wakeup_time = NA,
-      mean_time_in_bed = NA
+      mean_time_in_bed = NA,
+      sleep_efficiency = NA
     ))
   }
   col <- get_session_colnames(sessions, col_names)
@@ -30,7 +32,12 @@ get_sessions_summary <- function(sessions, col_names = NULL) {
       total_sessions = dplyr::n(),
       mean_sleep_onset = if (!is.null(col$time_at_sleep_onset)) mean_time(.data[[col$time_at_sleep_onset]]) else NA,
       mean_wakeup_time = if (!is.null(col$time_at_wakeup)) mean_time(.data[[col$time_at_wakeup]]) else NA,
-      mean_time_in_bed = if (!is.null(col$time_in_bed)) mean(.data[[col$time_in_bed]]) / 3600 else NA
+      mean_time_in_bed = if (!is.null(col$time_in_bed)) mean(.data[[col$time_in_bed]]) / 3600 else NA,
+      sleep_efficiency = if (!is.null(col$time_in_bed) && !is.null(col$sleep_period)) {
+        paste0(as.integer(round(mean(sessions[[col$sleep_period]], na.rm = TRUE) / mean(sessions[[col$time_in_bed]], na.rm = TRUE) * 100)), "%")
+      } else {
+        NA
+      },
     )
 
   if ("annotation" %in% colnames(sessions)) {